+++ src/core/server/player/index.ts
@@ -7,5 +7,6 @@ export * as permission from '@AthenaServer/player/permissions';
 export * as safe from '@AthenaServer/player/safe';
 export * as set from '@AthenaServer/player/setter';
 export * as sync from '@AthenaServer/player/sync';
+export * as teleport from '@AthenaServer/player/teleport';
 export * as toolbar from '@AthenaServer/player/toolbar';
 export * as weapons from '@AthenaServer/player/weapons';